moving result list options interface back into the ResultList file
https://coveord.atlassian.net/browse/JSUI-2237
